added stack folder

+++ sorting/course_allocation.c
@@ -1,27 +1,27 @@
 #include<stdio.h>
 #include<stdlib.h>
 
-void merge(int* arr1,int*arr2, int* arr3,int l,int mid,int r){
-    int* left1=malloc(sizeof(int)*(mid-l+1));
-    int* left2=malloc(sizeof(int)*(mid-l+1));
-    int* left3=malloc(sizeof(int)*(mid-l+1));
-    for(int i=l;i<=mid;i++){
+void merge(long long* arr1,long long*arr2, long long* arr3,long long l,long long mid,long long r){
+    long long* left1=malloc(sizeof(long long)*(mid-l+1));
+    long long* left2=malloc(sizeof(long long)*(mid-l+1));
+    long long* left3=malloc(sizeof(long long)*(mid-l+1));
+    for(long long i=l;i<=mid;i++){
         // printf("_inl\n");
         left1[i-l] = arr1[i];
         left2[i-l] = arr2[i];
         left3[i-l] = arr3[i];
     }
-    int* right1=malloc(sizeof(int)*(r-mid));
-    int* right2=malloc(sizeof(int)*(r-mid));
-    int* right3=malloc(sizeof(int)*(r-mid));
-    for(int i=mid+1;i<=r;i++){
+    long long* right1=malloc(sizeof(long long)*(r-mid));
+    long long* right2=malloc(sizeof(long long)*(r-mid));
+    long long* right3=malloc(sizeof(long long)*(r-mid));
+    for(long long i=mid+1;i<=r;i++){
         // printf("_inr\n");
         right1[i-mid-1] = arr1[i];
         right2[i-mid-1] = arr2[i];
         right3[i-mid-1] = arr3[i];
     }
-    int i=0,j=0;
-    for(int k=l;k<=r;k++){
+    long long i=0,j=0;
+    for(long long k=l;k<=r;k++){
         if(i>=(mid-l+1)){
             arr1[k] = right1[j] ;
             arr2[k] = right2[j] ;
@@ -45,16 +45,11 @@ void merge(int* arr1,int*arr2, int* arr3,int l,int mid,int r){
             arr3[k] = right3[j] ;j++;
         }
     }
-    // for(int i=0; i <= mid-l;i++)free(left1[i]);
-    // free(left1);
-    // for(int i=0; i <= mid-r-1;i++)free(right1[i]);
-    // free(right1);
-    // printf("%d*%d\n",l,r);
     return ;
 }
-void mergesort(int* arr1,int* arr2,int* arr3,int l,int r){
+void mergesort(long long* arr1,long long* arr2,long long* arr3,long long l,long long r){
     if(r-l<=0)return;
-    int mid=(r-l)/2 + l;
+    long long mid=(r-l)/2 + l;
     mergesort(arr1,arr2,arr3,l,mid);
     mergesort(arr1,arr2,arr3,mid+1,r);
     merge(arr1,arr2,arr3,l,mid,r);
@@ -62,52 +57,19 @@ void mergesort(int* arr1,int* arr2,int* arr3,int l,int r){
 }
 int main()
 {
-    int t;
-    scanf("%d",&t);
-    int arr0[t],arr1[t],arr2[t];
-    for(int i=0;i<t;i++){
-        int x,y;
-        scanf("%d %d",&x,&y);arr0[i]=x-y;
+    long long t;
+    scanf("%lld",&t);
+    long long arr0[t],arr1[t],arr2[t];
+    for(long long i=0;i<t;i++){
+        long long x,y;
+        scanf("%lld %lld",&x,&y);arr0[i]=x-y;
         arr1[i]=x;arr2[i]=y;
-        // printf("_in1\n");
     }
-    // for(int j=0;j<(t);j++){
-    //     printf("%d ",arr1[j]);
-    // }
-    // printf("\n");
-    // for(int j=0;j<(t);j++){
-    //     printf("%d ",arr2[j]);
-    // }
-    // printf("\n");
-    // for(int j=0;j<(t);j++){
-    //     printf("%d ",arr0[j]);
-    // }
-    // printf("\n");
-
-    int sum = 0;
+    long long sum = 0;
     mergesort(arr1,arr2,arr0,0,t-1);
-
-    // for(int j=0;j<(t);j++){
-    //     printf("%d ",arr1[j]);
-    // }
-    // printf("\n");
-    // for(int j=0;j<(t);j++){
-    //     printf("%d ",arr2[j]);
-    // }
-    // printf("\n");
-    // for(int j=0;j<(t);j++){
-    //     printf("%d ",arr0[j]);
-    // }
-    // printf("\n");
-
-
-    for(int j=0;j<(t);j++){
+    for(long long j=0;j<(t);j++){
         if(j<t/2){sum+=arr1[j];}
         else sum+=arr2[j];
     }
-    
-    // printf("\n");printf("|%d| ",sum);
-
-    printf("%d\n",sum);
-    // printf("\n");
+    printf("%lld\n",sum);
 }
